added check for compilation '
